using lon in MST kernel

+++ src/generators/mst/main.c
@@ -74,10 +74,10 @@ static const char *pdfnames[NR_PDFS] = {
  */
 struct
 {
-	int nintervals;      /**< Number of sampling intervals.         */
-	int npoints;         /**< Number of points.                     */
+	long nintervals;      /**< Number of sampling intervals.         */
+	long npoints;         /**< Number of points.                     */
 	const char *pdfname; /**< Name of probability density function. */
-	int pdfid;           /**< ID of probability density function.   */
+	long pdfid;           /**< ID of probability density function.   */
 } args = {0, 0, NULL, 0};
  
 /**
@@ -106,10 +106,10 @@ static void usage(void)
  * @param argc Number of arguments.
  * @param argv Arguments.
  */
-static void readargs(int argc, const char **argv)
+static void readargs(long argc, const char **argv)
 {
 	/* Parse command line arguments. */
-	for (int i = 1; i < argc; i++)
+	for (long i = 1; i < argc; i++)
 	{
 		if (!strcmp(argv[i], "--nintervals"))
 			args.nintervals = atoi(argv[i + 1]);
@@ -128,7 +128,7 @@ static void readargs(int argc, const char **argv)
 		error("invalid number of tasks");
 	if (args.pdfname == NULL)
 		error("unsupported probability density function");
-	for (int i = 0; i < NR_PDFS; i++)
+	for (long i = 0; i < NR_PDFS; i++)
 	{
 		if (!strcmp(args.pdfname, pdfnames[i]))
 		{
@@ -188,7 +188,7 @@ int main(int argc, const char **argv)
 	
 	/* Dump input data. */
 	printf("%d\n", args.npoints);
-	for (int i = 0; i < args.npoints; i++)
+	for (long i = 0; i < args.npoints; i++)
 	{
 		double y;
